Do not pass DockCtx to docklet methods
The original idea was to have a place with all the "global" things like
various services' shared states (now no longer a thing, state is only passed via bus)
and objects that perform stuff (like the wayland seat to activate apps)
and so on.

But especially w/o the shared state, it just seemed like a pointless bloat object.
So now DockCtx is only a little substructure of the dock that contains subscribers
that should always be cloned and never used directly..