Fix dock stuck open on one monitor by making IcedInstance run() "drop-safe"
This isn't a common term but basically, some futures are okay to drop
(which we do when selecting on one and forgetting the rest), like bus subscribers,
but some are not. The (impl Future) generated by this async fn was actually not "drop-safe"
because the leave_timeout could be left empty, i.e. the "put it back" code would not be executed.